Move some things around and consolidate
