+++ src/assets/cube.blend
