+++ src/assets/cube.glb
