+++ src/assets/white.png
