+++ src/lib/data_connector/data_connector.py
@@ -6,7 +6,7 @@ import blosc
 from lib.data_provider.abstract_data_provider import AbstractDataProvider
 from lib.model.record import Record
 from lib.model.scan import Scan
-from lib.model.serialization.serialization import deserializeScan, serializeScan
+from lib.model.convert.serialization import deserializeScan, serializeScan
 
 
 class DataConnector:
